requirements.txt
numpy
matplotlib
psutil
cupy-cuda11x